Move ejscreen and airdata to us_epa and add __init__.py
