+++ scripts/us_epa/airdata/air_quality_aggregate_test.py
@@ -12,12 +12,12 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 '''
-Unit tests for air_quality.py
+Unit tests for air_quality_aggregate.py
 
-Usage: python3 air_quality_test.py
+Usage: python3 -m unittest discover -v -s ../ -p "air_quality_aggregate_test.py"
 '''
 import unittest, csv, os, tempfile
-from air_quality_aggregate import create_csv, write_csv, write_tmcf
+from .air_quality_aggregate import create_csv, write_csv, write_tmcf
 
 module_dir_ = os.path.dirname(__file__)
