+++ scripts/us_epa/airdata/air_quality_test.py
@@ -14,10 +14,10 @@
 '''
 Unit tests for air_quality.py
 
-Usage: python3 air_quality_test.py
+Usage: python3 -m unittest discover -v -s ../ -p "air_quality_test.py"
 '''
 import unittest, csv, os, tempfile
-from air_quality import create_csv, write_csv, write_tmcf
+from .air_quality import create_csv, write_csv, write_tmcf
 
 module_dir_ = os.path.dirname(__file__)
